refactor user routes

+++ users/router.py
@@ -9,7 +9,7 @@ from users.schemas import User, UserCreate, UserUpdate
 router = APIRouter()
 
 
-@router.get("/users", response_model=GetUsersResponse, include_in_schema=False)
+@router.get("/users", response_model=GetUsersResponse)
 def read_users(auth_user_id=Depends(get_auth_user_id)):
     if not isinstance(auth_user_id, int):
         return auth_user_id
@@ -21,37 +21,48 @@ def read_users(auth_user_id=Depends(get_auth_user_id)):
         raise HTTPException(status_code=500, detail="Validation error occurred")
 
 
-@router.post("/users/create", response_model=User, include_in_schema=False)
+@router.post("/users/create", response_model=UserSuccessResponse)
 def create_user(register_form: UserCreate, auth_user_id=Depends(get_auth_user_id)):
     if not isinstance(auth_user_id, int):
         return auth_user_id
     db_user = crud.get_user_by_email(email=register_form.email)
-    db_account = crud.get_account(account_id=register_form.company_id)
+    register_form = register_form.dict()
+    register_form['company_id'] = crud.get_user(auth_user_id)['company_id']
     if db_user:
         raise HTTPException(status_code=400, detail="Email already registered")
-
-    if not db_account:
-        raise HTTPException(status_code=400, detail="Account not registered")
 
     return crud.create_user(new_user=register_form)
 
 
-@router.get("/users/{user_id}", response_model=User, include_in_schema=False)
+@router.get("/users/{user_id}", response_model=UserSuccessResponse)
 def read_user(user_id: int):
-    return crud.get_user(user_id=user_id)
+    return success_response({'user': crud.get_user(user_id=user_id)})
 
 
-@router.patch("/users/{user_id}", response_model=User, include_in_schema=False)
-def update_user(user_id: int, updated_data: UserUpdate):
+@router.patch("/users/{user_id}", response_model=StringResponse)
+def update_user(user_id: int, updated_data: UserUpdate, auth_user_id=Depends(get_auth_user_id)):
+    if not isinstance(auth_user_id, int):
+        return auth_user_id
+    auth_user = crud.get_user(auth_user_id)
+    required_user = crud.get_user(user_id)
+    if auth_user['company_id'] != required_user['company_id']:
+        return error_response('User is not from your company. Access denied', 404)
+
     db_user = crud.update_user(user_id=user_id, updated_data=updated_data)
     if not db_user:
         raise HTTPException(status_code=404, detail="User not found")
-    return db_user
+    return success_response('updated')
 
 
-@router.delete("/users/{user_id}", include_in_schema=False)
-def delete_user(user_id: int):
+@router.delete("/users/{user_id}", response_model=StringResponse)
+def delete_user(user_id: int, auth_user_id=Depends(get_auth_user_id)):
+    if not isinstance(auth_user_id, int):
+        return auth_user_id
+    auth_user = crud.get_user(auth_user_id)
+    required_user = crud.get_user(user_id)
+    if auth_user['company_id'] != required_user['company_id']:
+        return error_response('User is not from your company. Access denied', 404)
     deleted = crud.delete_user(user_id=user_id)
     if not deleted:
         raise HTTPException(status_code=404, detail="User not found")
-    return {"status": 200, "message": "User deleted"}
+    return success_response('deleted')
